Crate rs3cache

RS3 cache tool

Tools and api for reading and interpreting the RuneScape 3 game cache.

Installing the command line tool.

Install the Rust compiler.

Install the tool with

 cargo +nightly install --git https://github.com/mejrs/rs3cache/ rs3 --features=rs3

Usage

rs3 --help to see a list of commands:

USAGE:
    rs3.exe [OPTIONS]

OPTIONS:
        --assert-coherence      Checks whether the cache is in a consistent state. Indices 14, 40,
                                54, 55 are not necessarily complete
        --dump <DUMP>...        Dumps the given archives [possible values: all, configs, music,
                                achievements, sprites, locations, locations_each, tiles_each,
                                location_configs, location_configs_each, npc_config, item_configs,
                                maplabels, worldmaps, varbit_configs, structs, enums, underlays,
                                overlays]
    -h, --help                  Print help information
        --input <INPUT>         The path where to look for the current cache [env:
                                RS3_CACHE_INPUT_FOLDER=C:\ProgramData\Jagex\RuneScape] [default: ]
        --output <OUTPUT>       The path where to place output [env: RS3_CACHE_OUTPUT_FOLDER=]
                                [default: ]
        --render <RENDER>...    This exports them as small tiles, formatted as
                                `<layer>/<mapid>/<zoom>/<plane>_<x>_<y>.png`, suitable for use with
                                interactive map libraries such as <https://leafletjs.com/>, as seen
                                on <https://mejrs.github.io/> [possible values: all, map]

Building as a Python library.

Using maturin

Clone the repository:

git clone https://github.com/mejrs/rs3cache

Install Python, version 3.7 or newer.

Check that pip is installed (python -m pip --version).

Install maturin: pip install maturin.

Install the Rust compiler.

Navigate to the /rs3_py folder and build a Python wheel with

maturin build

Using the wheel from above, run

pip install <path to wheel>

For examples, see rs3cache/examples.

Using setuptools-rust

The following instructions are only tested on windows. If you run into issues on other platforms, try following here or build with maturin instead.

Clone the repository:

git clone https://github.com/mejrs/rs3cache

Install Python, version 3.7 or newer..

Check that pip is installed (python -m pip --version).

Install setuptools: pip install setuptools.

Install setuptools-rust: pip install setuptools-rust.

Install the Rust compiler.

Navigate to the /rs3_py folder and build a Python wheel with

python setup.py install

For examples, see rs3cache/examples.

Modules

clirs3 or osrs or legacy

definitionsrs3 or osrs or legacy

Entities that can be deserialized from cache data.

ffirs3 or osrs or legacy

Foreign function interfaces to rs3cache.

impl_

renderersNon-WebAssembly and (rs3 or osrs or legacy)

Functions for rendering the map.

structuresrs3 or osrs or legacy

Contains structures that are used in multiple different configs.

typesrs3 or osrs or legacy

Various data types