Module rs3cache::cache::index

The interface between rs3cache and the cache database.

Structs

CacheIndex
Container of Archives.

Initial
Initial state of CacheIndex.

IntoIter
Iterator over all Archives of self. Yields in arbitrary order.

IntoIterGrouped
An iterator of ArchiveGroups. Used internally by renderers that need to know about the surrounding GroupMapSquare.

Constants

SYS_VAR
System variable for the location of the cache folder.

Functions

assert_coherence
Asserts whether all indices’ metadata match their contents. Indices 14, 40, 54, 55 are not necessarily complete.