Struct rs3cache::cache::index::CacheIndex

pub struct CacheIndex<S: IndexState> { /* fields omitted */ }

Container of Archives.

Implementations

impl<S> CacheIndex<S> where
S: IndexState,

pub const fn index_id(&self) -> u32

The index id of self, corresponding to the raw/js5-{index_id}.jcache file being held.

pub const fn metadatas(&self) -> &IndexMetadata

Returns a view over the IndexMetadata of self.

pub fn archive(&self, archive_id: u32) -> CacheResult<Archive>

Get an Archive from self.

Errors

Raises ArchiveNotFoundError if archive_id is not in self.

pub fn assert_coherence(&self) -> CacheResult<()>

Assert whether the cache held by self is in a coherent state.

Errors

May raise CrcError, VersionError or ArchiveNotFoundError if the cache is not in a logical state.

Notes

Indices VORBIS, AUDIOSTREAMS, TEXTURES_PNG_MIPPED and TEXTURES_ETC tend to never complete. For these, simply ignore ArchiveNotFoundError.

impl CacheIndex<Initial>

pub fn new(index_id: u32) -> CacheResult<CacheIndex<Initial>>

Constructor for CacheIndex.

Errors

Raises CacheNotFoundError if the cache database cannot be found.

pub fn retain(self, ids: Vec<u32>) -> CacheIndex<Truncated>

Retain only those archives that are in ids. Advances self to the Truncated state.

Panics

Panics if any of ids is not in self.

pub fn grouped(
self,
dim_i: RangeInclusive<i32>,
dim_j: RangeInclusive<i32>
) -> CacheIndex<Grouped>

Groups archives according to their surface proximity. Only valid for the MAPSV2 index. Advances self to the Grouped state.

Panics

Panics if self.index_id() != IndexType::MAPSV2.

pub fn try_clone(&self) -> CacheResult<Self>

Attempt to clone self.

impl CacheIndex<Truncated>

pub fn grouped(
self,
dim_i: RangeInclusive<i32>,
dim_j: RangeInclusive<i32>
) -> CacheIndex<TruncatedGrouped>

Groups archives according to their surface proximity. Only valid for the MAPSV2 index. Advances self to the TruncatedGrouped state.

Panics

Panics if self.index_id() != IndexType::MAPSV2.