Struct rs3cache::definitions::indextype::MapFileType

pub struct MapFileType;

This is supported on crate features rs3 or osrs or legacy only.

Enumeration of the files in the MAPSV2 archives.

Implementations

impl MapFileType

pub const LOCATIONS: u32

This is supported on crate features rs3 or 2013_4_shim only.

Deserializes to the sequence of Locations in self.

pub const WATER_LOCATIONS: u32

This is supported on crate features rs3 or 2013_4_shim only.

Deserializes to a sequence of underwater Locations in self. Not implemented.

pub const NPCS: u32

This is supported on crate features rs3 or 2013_4_shim only.

Deserializes to a sequence of all npcs in self. Only mapsquares which used to have a “zoom around” login screen, or are derived from one that had, have this file.

pub const TILES: u32

This is supported on crate features rs3 or 2013_4_shim only.

Deserializes to the Tiles of self.

pub const WATER_TILES: u32

This is supported on crate features rs3 or 2013_4_shim only.

Deserializes to the underwater Tiles of self.

pub const UNKNOWN_5: u32

This is supported on crate features rs3 or 2013_4_shim only.

pub const UNKNOWN_6: u32

This is supported on crate features rs3 or 2013_4_shim only.

pub const UNKNOWN_7: u32

This is supported on crate features rs3 or 2013_4_shim only.

pub const UNKNOWN_8: u32

This is supported on crate features rs3 or 2013_4_shim only.

pub const UNKNOWN_9: u32

This is supported on crate features rs3 or 2013_4_shim only.