Module rs3cache::definitions::location_configs

Configuration of game locations.

Structs

ColourReplacements

ExtendedLocationMorphTable
Like LocationMorphTable, but with a default value.

HeadModels

LocationConfig
Describes the properties of a given Location.

LocationMorphTable
Contains an array of possible ids this location can morph into, controlled by either a varbit or varp.

Models

OldCursors

RecolourPalette

Textures

Unknown78

Unknown79

Unknown160

Unknown163

Unknown173

Unknown181

Unknown201

Functions

export
Save the location configs as location_configs.json. Exposed as --dump location_configs.

export_each
Save the location configs as individual json files.