Module rs3cache::definitions::location_configs::location_config_fields_impl

Defines the structs used as fields of LocationConfig,

Structs

ColourReplacements

ExtendedLocationMorphTable
Like LocationMorphTable, but with a default value.

HeadModels

LocationMorphTable
Contains an array of possible ids this location can morph into, controlled by either a varbit or varp.

Models

OldCursors

RecolourPalette

Textures

Unknown78

Unknown79

Unknown160

Unknown163

Unknown173

Unknown181

Unknown201