Struct rs3cache::definitions::location_configs::LocationConfig

pub struct LocationConfig {

    pub id: u32,
    pub models: Option<Models>,
    pub name: Option<String>,
    pub dim_x: Option<u8>,
    pub dim_y: Option<u8>,
    pub unknown_17: Option<bool>,
    pub is_transparent: Option<bool>,
    pub unknown_19: Option<u8>,
    pub unknown_21: Option<bool>,
    pub unknown_22: Option<bool>,
    pub occludes_1: Option<bool>,
    pub unknown_24: Option<u32>,
    pub unknown_27: Option<bool>,
    pub unknown_28: Option<u8>,
    pub ambient: Option<i8>,
    pub actions: Option<[Option<String>; 5]>,
    pub contrast: Option<i8>,
    pub colour_replacements: Option<ColourReplacements>,
    pub textures: Option<Textures>,
    pub recolour_palette: Option<Vec<(u16, u16)>>,
    pub unknown_44: Option<u16>,
    pub unknown_45: Option<u16>,
    pub mirror: Option<bool>,
    pub model: Option<bool>,
    pub scale_x: Option<u16>,
    pub scale_y: Option<u16>,
    pub scale_z: Option<u16>,
    pub unknown_69: Option<u8>,
    pub translate_x: Option<u16>,
    pub translate_y: Option<u16>,
    pub translate_z: Option<u16>,
    pub unknown_73: Option<bool>,
    pub blocks_ranged: Option<bool>,
    pub unknown_75: Option<u8>,
    pub morphs_1: Option<LocationMorphTable>,
    pub unknown_78: Option<Unknown78>,
    pub unknown_79: Option<Unknown79>,
    pub unknown_81: Option<u8>,
    pub unknown_82: Option<bool>,
    pub unknown_88: Option<bool>,
    pub unknown_89: Option<bool>,
    pub is_members: Option<bool>,
    pub morphs_2: Option<ExtendedLocationMorphTable>,
    pub unknown_93: Option<u16>,
    pub unknown_94: Option<bool>,
    pub unknown_95: Option<u16>,
    pub unknown_96: Option<bool>,
    pub unknown_97: Option<bool>,
    pub unknown_98: Option<bool>,
    pub unknown_99: Option<()>,
    pub unknown_101: Option<u8>,
    pub mapscene: Option<u16>,
    pub occludes_2: Option<bool>,
    pub unknown_104: Option<u8>,
    pub headmodels: Option<HeadModels>,
    pub mapfunction: Option<u16>,
    pub member_actions: Option<[Option<String>; 5]>,
    pub unknown_160: Option<Unknown160>,
    pub unknown_162: Option<i32>,
    pub unknown_163: Option<Unknown163>,
    pub unknown_164: Option<u16>,
    pub unknown_165: Option<u16>,
    pub unknown_166: Option<u16>,
    pub unknown_167: Option<u16>,
    pub unknown_168: Option<bool>,
    pub unknown_169: Option<bool>,
    pub unknown_170: Option<u16>,
    pub unknown_171: Option<u16>,
    pub unknown_173: Option<Unknown173>,
    pub unknown_177: Option<bool>,
    pub unknown_178: Option<u8>,
    pub unknown_186: Option<u8>,
    pub unknown_188: Option<bool>,
    pub unknown_189: Option<bool>,
    pub cursors: Option<[Option<u16>; 6]>,
    pub unknown_196: Option<u8>,
    pub unknown_197: Option<u8>,
    pub unknown_198: Option<bool>,
    pub unknown_199: Option<bool>,
    pub unknown_200: Option<bool>,
    pub unknown_201: Option<Unknown201>,
    pub unknown_202: Option<u16>,
    pub params: Option<ParamTable>,

}

Describes the properties of a given Location.

Fields
id: u32

Its id.

models: Option<Models>

A mapping of possible types to models.

name: Option<String>

Its name, if present.

dim_x: Option<u8>

Its west-east dimension, defaulting to 1 if not present.

Code using this value must account for the location’s rotation.

dim_y: Option<u8>

Its south-north dimension, defaulting to 1 if not present.

Code using this value must account for the location’s rotation.

unknown_17: Option<bool>is_transparent: Option<bool>unknown_19: Option<u8>

Flag for whether this object has a red rather than a white line on the map.

unknown_21: Option<bool>unknown_22: Option<bool>occludes_1: Option<bool>unknown_24: Option<u32>unknown_27: Option<bool>unknown_28: Option<u8>ambient: Option<i8>actions: Option<[Option<String>; 5]>

What rightclick options this location has, if any.

contrast: Option<i8>colour_replacements: Option<ColourReplacements>textures: Option<Textures>recolour_palette: Option<Vec<(u16, u16)>>unknown_44: Option<u16>unknown_45: Option<u16>mirror: Option<bool>model: Option<bool>scale_x: Option<u16>scale_y: Option<u16>scale_z: Option<u16>unknown_69: Option<u8>translate_x: Option<u16>translate_y: Option<u16>translate_z: Option<u16>unknown_73: Option<bool>blocks_ranged: Option<bool>

Whether this location can be interacted through with e.g. ranged/magic combat, telegrab etc.

unknown_75: Option<u8>morphs_1: Option<LocationMorphTable>

This location can have different appearances depending on a player’s varp/varbits.

unknown_78: Option<Unknown78>unknown_79: Option<Unknown79>unknown_81: Option<u8>unknown_82: Option<bool>unknown_88: Option<bool>unknown_89: Option<bool>is_members: Option<bool>morphs_2: Option<ExtendedLocationMorphTable>

This location can have different appearances depending on a players varbits, like the morphs_1 field, but with a default value.

unknown_93: Option<u16>unknown_94: Option<bool>unknown_95: Option<u16>unknown_96: Option<bool>unknown_97: Option<bool>unknown_98: Option<bool>unknown_99: Option<()>unknown_101: Option<u8>mapscene: Option<u16>

Reference to a MapScene that is drawn on the map.

occludes_2: Option<bool>unknown_104: Option<u8>headmodels: Option<HeadModels>mapfunction: Option<u16>member_actions: Option<[Option<String>; 5]>unknown_160: Option<Unknown160>unknown_162: Option<i32>unknown_163: Option<Unknown163>unknown_164: Option<u16>unknown_165: Option<u16>unknown_166: Option<u16>unknown_167: Option<u16>unknown_168: Option<bool>unknown_169: Option<bool>unknown_170: Option<u16>unknown_171: Option<u16>unknown_173: Option<Unknown173>unknown_177: Option<bool>unknown_178: Option<u8>unknown_186: Option<u8>unknown_188: Option<bool>unknown_189: Option<bool>cursors: Option<[Option<u16>; 6]>unknown_196: Option<u8>unknown_197: Option<u8>unknown_198: Option<bool>unknown_199: Option<bool>unknown_200: Option<bool>unknown_201: Option<Unknown201>unknown_202: Option<u16>params: Option<ParamTable>
Implementations

impl LocationConfig

pub fn dump_all(config: &Config) -> CacheResult<BTreeMap<u32, Self>>

Returns a mapping of all location configurations