Module rs3cache::definitions::mapsquares

Data comprising the game surface.

The game map is divided in (potentially) 128 by 256 MapSquares, though this seems to be limited to 100 by 200. These coordinates are referred to as i and j.

Each MapSquare itself is comprised of 64 by 64 Tiles. These coordinates are referred to as x and y. They have four elevations, referred to as p or plane.

Structs

GroupMapSquare

A group of adjacent MapSquares.

GroupMapSquareIterator

Iterates over GroupMapSquare in arbitrary order.

MapSquare

Represents a section of the game map

MapSquareIterator

Iterates over all MapSquares in arbitrary order. Iterates over all MapSquares in arbitrary order.

MapSquares

Functions

export_locations_by_id

Saves all occurences of every object id as a json file to the folder out/data/rs3/locations.

export_locations_by_square

Saves all occurences of every object id as a json file to the folder out/data/rs3/locations.

Type Definitions

ColumnIter

Iterator over a columns of planes with their x, y coordinates