Module rs3cache::definitions::npc_configs::npc_config_fields

Defines the structs used as fields of NpcConfig,

Structs

AmbientSounds

ColourReplacements

ExtendedNpcMorphTable

Like NpcMorphTable, but with a default value.

HeadModels

NpcModels

NpcMorphTable

Contains an array of possible ids this npc can morph into, controlled by either a varbit or varp.

OldCursors

Quests

RecolourPalette

Shadow

ShadowIntensity

Textures

Translations

Unknown155

Unknown164

Unknown179