Function rs3cache::definitions::sprites::dumps

pub fn dumps(
    scale: u32,
    ids: Vec<u32>,
    config: &Config
) -> CacheResult<BTreeMap<(u32, u32), Sprite>>

Available on crate features rs3 or osrs or legacy only.

Returns a BTreeMap holding all sprites in ids.

Sprites are scaled according to scale, which may not be 0.

Errors

Raises [CacheError](rs3cache_backend::error::{CacheError, Context}) if any of ids does not correspond to a sprite.

Panics

Panics if scale == 0.