Module rs3cache::ffi::python

Python bindings for rs3cache.

Setup

You need to build binary wheels locally. The following instructions are only guaranteed on windows. If you run into issues on other platforms, please follow here or try maturin instead.

git clone https://github.com/mejrs/rs3cache.

Install Python, version 3.6 or newer.

Check that pip is installed (python -m pip --version).

Install setuptools: pip install setuptools.

Install setuptools-rust: pip install setuptools-rust.

Install the Rust compiler.

Configure rustup to use the nightly version: rustup default nightly.

Navigate to this repository and run python setup.py install.

Either:

Create a system variable named RUNESCAPE_CACHE_FOLDER and set its value to where your cache is located. Typically, this is %ProgramData%\Jagex\RuneScape.

Copy the entire cache and place it in the raw folder.

Usage

All the following can be imported as:

from rs3cache import *

Structs

PyCacheIndex
Wrapper over CacheIndex. The Python alias for this is Index

PyCacheIndexIter
Iterator over all archives in an Index.

PyMapSquare
Obtained from PyMapSquares’s get method.

PyMapSquares
Container of PyMapSquares. Accessible with from rs3cache import MapSquares.

Functions

PyInit_rs3cache⚠
This autogenerated function is called by the python interpreter when importing the module.

get_location_configs
Wrapper for LocationConfig::dump_all

get_npc_configs
Wrapper for NpcConfig::dump_all

get_varbit_configs
Wrapper for VarbitConfig::dump_all