Crate rs3cache

RS3 cache tool

Tools and api for reading and interpreting the RuneScape 3 game cache.

Installing the command line tool.

Install the Rust compiler.

Install the tool with

 cargo +nightly install --git https://github.com/mejrs/rs3cache/ rs3cache --features=rs3

Usage

rs3cache --help to see a list of commands:

USAGE:
    rs3cache.exe [FLAGS] [OPTIONS]

FLAGS:
        --assert-coherence
            Checks whether the cache is in a consistent state. Indices 14, 40, 54, 55 are not necessarily complete

    -h, --help
            Prints help information

    -V, --version
            Prints version information

OPTIONS:
        --dump <dump>...
            Allowed values: [all, sprites, locations, location_configs, location_configs_each, npc_configs, item_configs,
            maplabels, worldmaps, varbit_configs, structs, enums, underlays, overlays]

            Dumps the given archives.
        --input <input>
            The path where to look for the current cache. If omitted this falls back to the "RS3_CACHE_INPUT_FOLDER"
            environment variable and then to the current folder if not set [env: RS3_CACHE_INPUT_FOLDER=]  [default: ]
        --output <output>
            The path where to place output. If omitted this falls back to the "RS3_CACHE_OUTPUT_FOLDER" environment
            variable and then to the current folder if not set [env: RS3_CACHE_OUTPUT_FOLDER=]  [default: ]
        --render <render>...
            Allowed values: [all, map]

            This exports them as small tiles, formatted as `<layer>/<mapid>/<zoom>/<plane>_<x>_<y>.png`, suitable for
            use with interactive map libraries such as https://leafletjs.com/, as seen on https://mejrs.github.io/

Building as a Python library.

Using maturin

Clone the repository:

git clone https://github.com/mejrs/rs3cache

Install Python, version 3.7 or newer.

Check that pip is installed (python -m pip --version).

Install maturin: pip install maturin.

Install the Rust compiler.

Navigate to the /rs3_py folder and build a Python wheel with

maturin build

Using the wheel from above, run

pip install <path to wheel>

For examples, see rs3cache/examples.

Using setuptools-rust

The following instructions are only tested on windows. If you run into issues on other platforms, try following here or build with maturin instead.

Clone the repository:

git clone https://github.com/mejrs/rs3cache

Install Python, version 3.7 or newer..

Check that pip is installed (python -m pip --version).

Install setuptools: pip install setuptools.

Install setuptools-rust: pip install setuptools-rust.

Install the Rust compiler.

Navigate to the /rs3_py folder and build a Python wheel with

python setup.py install

For examples, see rs3cache/examples.

Re-exports

pub use rs3cache_backend as cache;

sqlite or dat2 or dat

Modules

clirs3 or osrs or legacy

definitionsrs3 or osrs or legacy

Entities that can be deserialized from cache data.

ffirs3 or osrs or legacy

Foreign function interfaces to rs3cache.

renderersNon-WebAssembly and (rs3 or osrs or legacy)

Functions for rendering the map.

structuresrs3 or osrs or legacy

Contains structures that are used in multiple different configs.

typesrs3 or osrs or legacy

Various data types