Module rs3cache::definitions

Available on crate features rs3 or osrs or legacy only.

Entities that can be deserialized from cache data.

Modules

achievementsrs3

Configurations of Achievements Describes the properties of Achievements.

dbrows

enums

Describes the properties of enums.

indextype

The names of various indexes and archives.

item_configs

Describes the properties of items.

location_configs

Configuration of game locations.

locations

Describes the id, position, type and rotation of game objects.

maplabel_configs

mapscenesrs3 or 2009_1_shim

Configuration of images drawn on the world map. Describes text, sprites and polygons drawn on the map.

mapsquares

Data comprising the game surface.

musicrs3

npc_configs

Configuration of npcs.

npcs

Describes the position and id of npcs.

overlaysrs3 or osrs

Describes the colours of tiles.

sprites

Images displayed by the game client.

structs

Describes the properties of structs.

tiles

Describes the properties of game surface tiles.

underlaysrs3 or osrs

Describes the ground colours of tiles.

varbit_configs

Bitpacked player variables.

worldmapsrs3