Module rs3cache::definitions

Entities that can be deserialized from cache data.

Modules

location_configs
Configuration of game locations.

locations
Describes the id, position, type and rotation of game objects.

maplabel_configs
Describes text, sprites and polygons drawn on the map.

mapscenes
Configuration of images drawn on the world map.

mapsquares
Data comprising the game surface.

npc_configs
Configuration of npcs.

npcs
Describes the position and id of npcs.

overlays
Describes the colours of tiles.

sprites
Images displayed by the game client.

tiles
Describes the properties of game surface tiles.

underlays
Describes the ground colours of tiles.

varbit_configs
Bitpacked player variables.

worldmaps