Module rs3cache::definitions::location_configs::location_config_fields

Available on crate features rs3 or osrs or legacy only.

Defines the structs used as fields of LocationConfig,

Structs

ColourReplacements

ExtendedLocationMorphTable

Like LocationMorphTable, but with a default value.

HeadModels

LocationMorphTable

Contains an array of possible ids this location can morph into, controlled by either a varbit or varp.

Models

Models2

Textures

Unknown78

Unknown79

Unknown160

Unknown163

Unknown173

Unknown201

Unknown204