Struct rs3cache::definitions::locations::Location

pub struct Location {
    pub plane: Watery,
    pub i: u8,
    pub j: u8,
    pub x: u8,
    pub y: u8,
    pub id: u32,
    pub type: u8,
    pub rotation: u8,
}

Available on crate features rs3 or osrs or legacy only.

A location, also referred to as an “object”.

Fields
plane: Watery

The plane a.k.a elevation.

It can have any value in the range 0..=3.

i: u8

The horizontal MapSquare coordinate.

It can have any value in the range 0..=100.

All operations on this value should use explicit wrapping arithmetic.

j: u8

The vertical MapSquare coordinate.

It can have any value in the range 0..=200.

All operations on this value should use explicit wrapping arithmetic.

x: u8

The horizontal coordinate inside its MapSquare.

It can have any value in the range 0..=63. Locations that are not 1x1 have their most western tile as this coordinate.

y: u8

The vertical coordinate inside its MapSquare.

It can have any value in the range 0..=63. Locations that are not 1x1 have their most southern tile as this coordinate.

id: u32

The id corresponding to its LocationConfig.

type: u8

The type of this location. The models field of its LocationConfig maps models to its type.

rotation: u8

Its rotation, also known as “orientation”.

Implementations

impl Location

pub fn dump(i: u8, j: u8, tiles: &TileArray, buffer: Bytes) -> Vec<Self>

Constructor for Location.