Struct rs3cache::definitions::mapsquares::GroupMapSquare

pub struct GroupMapSquare { /* fields omitted */ }

A group of adjacent MapSquares.

Necessary for operations that need to care about surrounding mapsquares.

Implementations

impl GroupMapSquare

pub fn core_i(&self) -> u8

The horizontal coordinate of the central MapSquare.

It can have any value in the range 0..100.

pub fn core_j(&self) -> u8

The vertical coordinate of the central MapSquare.

It can have any value in the range 0..200.

pub fn core(&self) -> &MapSquare

Returns a reference to the central MapSquare.

pub fn iter(&self) -> Iter<'_, (u8, u8), MapSquare>

Iterates over all MapSquares of self in arbitrary order.

pub fn get(&self, key: &(u8, u8)) -> Option<&MapSquare>

Returns a view over a specific MapSquare..

pub fn tiles_iter(
&self,
plane: usize,
x: usize,
y: usize,
interp: isize
) -> Box<dyn Iterator<Item = &Tile>>

Returns a view over all tiles within interp of the Tile at plane, x, y.

pub fn all_locations_iter(&self) -> Box<dyn Iterator<Item = &Location>>

Returns a view over all locations in all MapSquares of self in arbitrary order.