Struct rs3cache::definitions::mapsquares::MapSquare

pub struct MapSquare {
    pub members: Option<u64>,
    /* private fields */
}

Available on crate features rs3 or osrs or legacy only.

Represents a section of the game map

Fields
members: Option<u64>
Available on crate feature rs3 only.

Member status of chunks in bits.

Implementations

impl MapSquare

pub fn i(&self) -> u8

The horizontal MapSquare coordinate.

It can have any value in the range 0..=100.

pub fn j(&self) -> u8

The vertical MapSquare coordinate.

It can have any value in the range 0..=200.

pub fn indexed_columns(&self) -> Option<ColumnIter<'_>>

Iterator over a columns of planes with their x, y coordinates

pub fn tiles(&self) -> Option<&TileArray>

Returns a view over the tiles field, if present

pub fn take_tiles(self) -> Option<TileArray>

Returns a view over the tiles field, if present

pub fn locations(&self) -> Option<&[Location]>

Returns a view over the locations field, if present.

pub fn take_locations(self) -> Option<Vec<Location>>

Take its locations, consuming self.

pub fn water_locations(&self) -> Option<&Result<Vec<Location>, CacheError>>

Available on crate features rs3 or 2013_4_shim only.

Returns a view over the locations field, if present.