Function rs3cache::definitions::sprites::dumps

pub fn dumps(
    scale: u32, 
    ids: Vec<u32>, 
    config: &Config
) -> CacheResult<BTreeMap<(u32, u32), Sprite>>

Returns a BTreeMap holding all sprites in ids.

Sprites are scaled according to scale, which may not be 0.

Errors

Raises CacheError if any of ids does not correspond to a sprite.

Panics

Panics if scale == 0.