Module rs3cache::renderers::map

Exports map tiles.

Re-exports

pub use self::mapcore::render;

Modules

base

Renders the ground colours.

lines

Responsible for drawing lines - doors, fences, walls and so on.

lineshape

Describes the shape of lines drawn by the map renderer.

mapcore

Core of the render - this is where everything is put together.

mapscenes

Responsible for drawing MapScene.

tileshape

Describes the shape of overlays drawn by the map renderer.